requirements.txt
anyio==3.6.2
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
Flask==2.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
Pillow==9.4.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.14.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dotenv==1.0.0
python-telegram-bot==20.1
python3-xlib==0.15
pythonw==3.0.3
pytweening==1.0.4
pywhat==5.1.0
requests==2.28.2
rfc3986==1.5.0
rich==10.16.2
sniffio==1.3.0
soupsieve==2.4
types-requests==2.28.11.16
types-urllib3==1.26.25.8
urllib3==1.26.15
Werkzeug==2.2.3
wikipedia==1.4.0
